+++ CHANGES.md
@@ -4,6 +4,7 @@
 - MongoDB: Added `MongoDBFullLoadTranslator` and `MongoDBCrateDBConverter`
 - Zyp: Fixed execution of collection transformation
 - Zyp: Added software test and documentation about flattening lists
+- MongoDB: Use `bson` package to parse BSON CANONICAL representation
 
 ## 2024/09/10 v0.0.15
 - Added Zyp Treatments, a slightly tailored transformation subsystem